standardized literals as 32bit and settled a use of uint32 for all literal type size matters. Soz, can't think of a clearer way to explain.

+++ CheckerZ/Parallelism/Parallel.hpp
@@ -11,7 +11,7 @@
 **/
 
 namespace CheckerZ { namespace Parallelism {
-		template<class Type, const size_t Size>
+		template<class Type, const uint32 Size>
 		using parallel_arr = concurrency::array_view<Type, Size>;
 } }
